+++ CHANGELOG.md
@@ -23,7 +23,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 ### Changed
 
 ### Fixed
-
+-   Fixed memory leaks, caused by non-working PyObject, PythonException, DelecateManager->Dispatcher finalizers.
 -   Fixed secondary PythonEngine.Initialize call, all sensitive static variables now reseted.
     This is a hidden bug. Once python cleaning up enough memory, objects from previous engine run becomes corrupted.
 -   Fixed Visual Studio 2017 compat (#434) for setup.py